inventory_management/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-leaflet==0.31.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
